requirements.txt
# ============================================================================
# Requirements for BSEC IAQ Anomaly Detection with KAN
# Python Version Required: 3.9.x (CRITICAL for KAN compatibility)
# ============================================================================

# ============================================================================
# Core Web Framework
# ============================================================================
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart==0.0.6

# ============================================================================
# Deep Learning - PyTorch Stack
# ============================================================================
# PyTorch 2.1.2 is the recommended version for Python 3.9
# Compatible with CUDA 11.8 or 12.1 if GPU is available
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2

# ============================================================================
# Machine Learning & Data Science
# ============================================================================
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# ============================================================================
# Time Series Database
# ============================================================================
influxdb==5.3.1              # For InfluxDB 1.x (DataFrameClient)
influxdb-client==1.38.0        # For InfluxDB 2.x (InfluxDBClient)
# Note: Choose based on your InfluxDB server version

# ============================================================================
# Visualization
# ============================================================================
matplotlib==3.8.2
seaborn==0.13.1

# ============================================================================
# Utilities
# ============================================================================
pyyaml==6.0.3
requests==2.32.5
tqdm==4.67.1

# ============================================================================
# KAN (Kolmogorov-Arnold Networks)
# ============================================================================
# CRITICAL: efficient-kan MUST be installed separately AFTER other packages
# This library ONLY works with Python 3.9 or earlier
#
# Installation command:
# pip install git+https://github.com/Blealtan/efficient-kan.git
#
# Alternative (if above fails):
# git clone https://github.com/Blealtan/efficient-kan.git
# cd efficient-kan
# pip install -e .
# ============================================================================

# ============================================================================
# Field Mapping (Optional — fuzzy matching for CSV column headers)
# ============================================================================
# rapidfuzz>=3.0.0            # Optional: fuzzy field mapping (python -m iaq4j map-fields)

# ============================================================================
# Development Tools (Optional - uncomment to install)
# ============================================================================
pytest==7.4.4              # Testing framework
pytest-cov==4.1.0          # Code coverage
hypothesis>=6.100.0        # Property-based testing
# black==24.1.1              # Code formatter
# flake8==7.0.0              # Linter
# isort==5.13.2              # Import sorter
# mypy==1.8.0                # Type checker
# ipython==8.12.3            # Enhanced REPL
# jupyter==1.0.0             # Jupyter notebooks
# notebook==7.0.8            # Jupyter notebook server

# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
#
# STEP 1: Verify Python Version
# -----------------------------
# python --version
# # Must show: Python 3.9.x
# # If not, install Python 3.9 first (see below)
#
# STEP 2: Create Virtual Environment (RECOMMENDED)
# -------------------------------------------------
# python -m venv venv
# source venv/bin/activate          # Mac/Linux
# # OR
# venv\Scripts\activate             # Windows
#
# STEP 3: Upgrade pip
# -------------------
# pip install --upgrade pip setuptools wheel
#
# STEP 4: Install Requirements
# ----------------------------
# pip install -r requirements.txt
#
# STEP 5: Install KAN (CRITICAL - Separate Step)
# -----------------------------------------------
# pip install git+https://github.com/Blealtan/efficient-kan.git
#
# STEP 6: Verify Installation
# ---------------------------
# python -c "import torch; print(f'✓ PyTorch {torch.__version__}')"
# python -c "import efficient_kan; print('✓ KAN installed')"
# python -c "from app.models import IAQPredictor; print('✓ IAQPredictor available')"
#
# ============================================================================
# PYTHON 3.9 INSTALLATION GUIDE
# ============================================================================
#
# Option A: Using pyenv (Recommended for Mac/Linux)
# --------------------------------------------------
# # Install pyenv
# brew install pyenv                    # Mac
# # OR
# curl https://pyenv.run | bash        # Linux
#
# # Install Python 3.9.18
# pyenv install 3.9.18
# pyenv local 3.9.18                   # Set for current directory
# python --version                     # Verify
#
# Option B: Using conda
# ---------------------
# conda create -n kan_env python=3.9
# conda activate kan_env
# python --version                     # Verify
#
# Option C: Direct Download
# -------------------------
# Download from: https://www.python.org/downloads/release/python-3918/
# Install and add to PATH
#
# ============================================================================
# TROUBLESHOOTING
# ============================================================================
#
# Issue: "No matching distribution found for torch==2.1.2"
# Solution:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#   # For CUDA 11.8:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#
# Issue: "Package 'pykan' requires a different Python"
# Solution: You're not on Python 3.9. Check: python --version
#
# Issue: KAN installation fails with "No module named 'torch'"
# Solution: Install PyTorch first (Step 4), THEN install KAN (Step 5)
#
# Issue: "ImportError: cannot import name 'IAQPredictor'"
# Solution: Replace app/models.py with the fixed version provided
#
# Issue: Import errors for evaluate or feature_engineering
# Solution: Replace training/*.py files with fixed versions
#
# ============================================================================
# GPU SUPPORT (Optional)
# ============================================================================
#
# For NVIDIA GPU with CUDA 11.8:
# -------------------------------
# pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu118
#
# For NVIDIA GPU with CUDA 12.1:
# -------------------------------
# pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu121
#
# Verify GPU:
# -----------
# python -c "import torch; print(f'CUDA available: {torch.cuda.is_available()}')"
#
# ============================================================================
# QUICK START AFTER INSTALLATION
# ============================================================================
#
# 1. Start the server:
#    uvicorn app.main:app --reload --host 0.0.0.0 --port 8000
#
# 2. Check health:
#    curl http://localhost:8000/health
#
# 3. Train models (if you have InfluxDB data):
#    python train_models.py
#
# 4. Create dummy models for testing:
#    python create_dummy_models.py
#
# 5. Test predictions:
#    python test_client.py
#
# ============================================================================